fix: remove code block link empty space underlining
The fact that the `before` and `after` elements occupy `0.2em` width in conjunction with the `text-decoration underline` rule causes surrounding empty space to be underlined too, which can in turn cause misunderstandings say in [package names](https://www.electronjs.org/docs/tutorial/quick-start#main-and-renderer-processes) and give one the impression that they are wrapped in underscores, f.e `_ipcMain_` instead of `ipcMain`.

To clarify this, the horizontal spacing was moved to the `code` element instead.